+++ client/imports/ui/routes/course.routes.ts
@@ -3,15 +3,15 @@
 import { RouterConfig, CanActivate } from '@angular/router';
 
 // Course Routing
-import CourseView from './course.ts';
+import CourseView from '../pages/course/course.ts';
 import { CourseEnroll } from './course.enroll.ts';
 
 // Pages
-import { GradeList } from './gradelist.ts';
-import { LabList } from './lablist.ts';
-import { CourseDashboard } from './course_dashboard.ts';
-import { LabView } from './labview.ts';
-import { GradeView } from './gradeview.ts';
+import { GradeList } from '../pages/course/gradelist.ts';
+import { LabList } from '../pages/course/lablist.ts';
+import { CourseDashboard } from '../pages/course/course_dashboard.ts';
+import { LabView } from '../pages/course/labview.ts';
+import { GradeView } from '../pages/course/gradeview.ts';
 
 export const courseRoutes: RouterConfig = [
   {
@@ -22,7 +22,7 @@ export const courseRoutes: RouterConfig = [
       { path: '', component: CourseDashboard },
       { path: 'grades', component: GradeList },
       { path: 'labs', component: LabList },
-      { path: 'labs/lab', component: LabView },
+      { path: 'labs/:labid', component: LabView },
       { path: 'grades/grade', component: GradeView }
   //  { path: 'users', as: 'UserList', component: UserList },
   //  { path: 'user/:userid', as: 'UserView', component: UserView },
